# Optimize StaticProxyFactory GetProxy and GetFieldInterceptionProxy methods